Fix remove from physics during sync
Before, the calls to remove physics bodies and physics colliders were
placed inside a loop which iterates only over existing entities.  Items were
never deleted immediately.  On the next insert, it detected there was an
entry with the same ID, deleted it and printed a warning.

This commit moved the delete calls outside of the loop.  It iterates over the
deleted ID's and removes them all from the physics engine.